requirements.txt
gmpy2==2.0.9
Pillow==7.0.0
tqdm==4.45.0
